% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{init}
\alias{init}
\title{New Vault Init
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
init(vault_path = NULL, key_size = 8192)
}
\arguments{
\item{vault_path}{Path to vault folder}

\item{key_size}{Size for RSA private key (Default: 8192)}
}
\description{
Initialize new vault
}
