% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_credit-card.R
\name{item_card.credit}
\alias{item_card.credit}
\title{Item: Credit Card   \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
item_card.credit(title, number, ccv, expiry = NULL, owner = NULL)
}
\arguments{
\item{title}{Object title. Will be used as identifier.}

\item{number}{Credit Card Number, number or string)}

\item{ccv}{Credit Card CCV number, expected number or string.}

\item{expiry}{Credit Card Expiry Date, an array of numbers labeled with 'y'(ear) and 'm'(onth)}

\item{owner}{Credit Card owner, expected string or object of class 'person'.}
}
\value{
a \code{card.credit} object.
}
\description{
Creator for credit card object types
}
\examples{
cc <- item_card.credit("CC - My Bank",
                       1234567890123456,
                       123,
                       c(m=1, y=2023),
                       person("Jane", "Doe"))

# Prints CC object
cc

# Prints CC number
cc$number

# Prints CC ccv
cc$ccv

# Prints CC expiry
cc$expiry

# Prints CC owner
cc$owner

}
