% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_card-loyalty.R
\name{item_card.loyalty}
\alias{item_card.loyalty}
\title{Item: Loyalty Card   \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
item_card.loyalty(title, number, expiry = NULL)
}
\arguments{
\item{title}{Object title. Will be used as identifier.}

\item{number}{Loyalty Card Number, number or string)}

\item{expiry}{Loyalty Card Expiry Date, an array of numbers labeled as 'y'(ear) and 'm'(onth)}
}
\value{
a \code{card.loyalty} object.
}
\description{
Creator for loyalty card object types
}
\examples{
loyal <- item_card.loyalty("ABC Shop",
                           1234567890123456,
                           c(m=1, y=2023))

# Prints LC object
loyal

# Prints LC number
loyal$number

# Prints LC expiry
loyal$expiry

}
