% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_pass-simple.R
\name{item_pass.simple}
\alias{item_pass.simple}
\title{Item: Simple Password   \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
item_pass.simple(title, url = NULL, username = NULL, password = NULL)
}
\arguments{
\item{title}{Object title. Will be used as identifier and file name.}

\item{url}{URL to which it applies.}

\item{username}{Username to store.}

\item{password}{Password to store.}
}
\value{
a \code{pass.simple} object.
}
\description{
Creator for simple password object types
}
\examples{
pw <- item_pass.simple("test_title", "test_url", "test_user", "my secret pass")
pw

}
