% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safevault-package.R
\docType{package}
\name{safevault-package}
\alias{safevault}
\alias{safevault-package}
\title{safevault: Secure Data Storage}
\description{
Manage secure storage of data.
}
\author{
\strong{Maintainer}: Bruno Gonçalves \email{bpvgoncalves@users.noreply.github.com} (\href{https://orcid.org/0000-0002-0797-7717}{ORCID})

}
\keyword{internal}
