% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_init}
\alias{vault_init}
\title{New Vault Init   \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
vault_init(vault_path = NULL, key_size = 8192)
}
\arguments{
\item{vault_path}{Path to vault file}

\item{key_size}{Size for RSA key (Default: 8192 bits)}
}
\description{
Initialize new vault
}
